Fixed overwritten #1749
Former-commit-id: 81545746424f4e0679cdd5870235620bb3805ab2

+++ src/com/liferay/util/FileUtil.java
@@ -376,8 +376,15 @@ public class FileUtil {
 
 		destination.delete();
 
-		copyFile(source, destination);
-		return source.delete();
+		boolean success = source.renameTo(destination);
+		
+		// if the rename fails, copy
+
+		if (!success) {
+			copyFile(source, destination);
+			success = source.delete();
+		}
+		return success;
 	}
 
 	public static String read(String fileName) throws IOException {
